Port BumpTop Mac to modern macOS (arm64): builds against Ogre 14, Qt 6, Bullet 3, protobuf 35
- New CMake/Ninja build (trunk/mac/CMakeLists.txt) replacing the Xcode 3 project
- Ogre 1.7 -> 14.4: static link, OverlaySystem registration, nodeless camera
  positioning, GLPlugin/STBI codec installed in code, affine-matrix and
  overlay-iterator API updates; render-order patch replicated in vendored Ogre
- Qt 4 -> 6: Widgets/Core5Compat split, QHttp -> QNetworkAccessManager, qSort
  shim, moc via AUTOMOC (dropped checked-in moc includes)
- Carbon removal: Process Manager -> NSRunningApplication, theme cursors ->
  NSCursor, Carbon Menu Manager context menu rewritten with NSMenu
- Sparkle/CMCrashReporter/LetsMove stubbed (PortShims/); dead QtWebKit theme
  browser and pre-10.6 QuickLook path excluded
- protobuf 2.4 -> 35: regenerated AllMessages.pb.*, un-final'd generated
  classes (legacy code inherits from them)

+++ trunk/mac/BumpTop/KeyboardEventManager.cpp
@@ -213,21 +213,21 @@ void KeyboardEventManager::keyDown(KeyboardEvent keyboard_event) {
     }
 
     // Arrow Keys
-    if (keyboard_event.characters.startsWith(QChar(ARROW_UP))) {
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_UP))) {
       bumptop_->scene()->room()->process_arrow_key(ARROW_UP, keyboard_event.modifier_flags);
     }
-    if (keyboard_event.characters.startsWith(QChar(ARROW_DOWN))) {
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_DOWN))) {
       bumptop_->scene()->room()->process_arrow_key(ARROW_DOWN, keyboard_event.modifier_flags);
     }
   } else {
     // Arrow Keys
-    if (keyboard_event.characters.startsWith(QChar(ARROW_UP)))
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_UP)))
       bumptop_->scene()->room()->process_arrow_key(ARROW_UP, keyboard_event.modifier_flags);
-    if (keyboard_event.characters.startsWith(QChar(ARROW_DOWN)))
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_DOWN)))
       bumptop_->scene()->room()->process_arrow_key(ARROW_DOWN, keyboard_event.modifier_flags);
-    if (keyboard_event.characters.startsWith(QChar(ARROW_LEFT)))
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_LEFT)))
       bumptop_->scene()->room()->process_arrow_key(ARROW_LEFT, keyboard_event.modifier_flags);
-    if (keyboard_event.characters.startsWith(QChar(ARROW_RIGHT)))
+    if (keyboard_event.characters.startsWith(QChar((ushort)ARROW_RIGHT)))
       bumptop_->scene()->room()->process_arrow_key(ARROW_RIGHT, keyboard_event.modifier_flags);
 
     // Close View
@@ -278,4 +278,3 @@ bool KeyboardEventManager::option_key_down() {
   return option_key_down_;
 }
 
-#include "BumpTop/moc/moc_KeyboardEventManager.cpp"
